# [libc++] Optimize {map,set}::insert(InputIterator, InputIterator)